backend/requirements.txt
alembic==1.14.1
click==8.1.8
ecdsa==0.19.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
pyasn1==0.6.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
typing_extensions==4.12.2
Werkzeug==3.1.3
SQLAlchemy==2.0.38
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
pytest==8.3.5
talisman==0.1.0
pytest-cov==6.0.0
Flask-SocketIO==5.5.1
eventlet==0.39.1
requests
azure-storage-blob